pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supplychain-app"
version = "1.6.2"
description = "SupplyChainApp (API Flask + frontend statique)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastexcel>=0.16.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "beautifulsoup4>=4.12.0",
    "html5lib>=1.1",
    "lxml>=5.2.0",
    "pandas>=2.2.0",
    "polars-lts-cpu==1.31.0",
    "pyarrow>=22.0.0",
    "requests>=2.32.5",
    "unidecode>=1.4.0",
    "xlsxwriter>=3.2.0",
    "pywin32>=311",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.17.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
